requirements.txt
Django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.3
PyJWT>=2.9
python-dotenv>=1.0
argon2-cffi>=23.1
psycopg2-binary>=2.9
